requirements.txt
appdirs==1.4.3
click==6.7
clickclick==1.2.1
connexion==1.0.129
constantly==15.1.0
Flask==0.12.1
Flask-Cors==3.0.2
Flask-Twisted==0.1.2
incremental==16.10.1
itsdangerous==0.24
Jinja2==2.9.6
jsonschema==2.6.0
MarkupSafe==1.0
observable==0.3.2
packaging==16.8
psycopg2==2.7.1
py==1.4.33
pycrypto==2.6.1
pyparsing==2.2.0
pytest==3.0.7
pytest-cov==2.5.1
python-dateutil==2.6.0
PyYAML==3.12
requests==2.14.2
six==1.10.0
SQLAlchemy==1.1.10
SQLAlchemy-Utils==0.32.14
strict-rfc3339==0.7
swagger-spec-validator==2.1.0
Twisted==16.6.0
Werkzeug==0.12.1
zope.interface==4.4.0
